+++ CHANGES.rst
@@ -5,7 +5,7 @@ Change log for sw.allotmentclub
 4.1.0 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Fix radicale server startup.
 
 
 4.0.1 (2018-02-23)